+++ scripts/asset-report.js
@@ -427,7 +427,7 @@ export default class AssetReport extends FormApplication {
    * @return {Promise<Response>}
    */
   static async FetchWithTimeout(resource, options) {
-    let timeout = game.settings.get(CONSTANTS.MODULE_NAME, CONSTANTS.SETTING_ASSET_TIMEOUT) || 60;
+    let timeout = game.settings.get(CONSTANTS.MODULE_NAME, CONSTANTS.SETTING_ASSET_TIMEOUT) ?? 60;
     if (options?.timeout) {
       timeout = options.timeout;
     }